Installation Process
An installation process has been created to install the basic environment
neded by the system in order to operate. This mostly consists of database
requirements that will be needed by Kore9.

The Installation uses a settings configuration found in data/base_install
which maps what parts have been installed and which have not. Later, updates
will be able to be downloaded and automatically installed.

The workflow of the exec has been updated to allow the installation to run
prior to attempting a connection to the database.

The databse was moved to 105 and the installation take the 101 spot.