Authenitcation and Session Handling
Authentication now works with the database using the tables class
object which extends the dbConnection class. This is used by the
authority class to query user account and group information to validate
login credentials.

The session was updated to store the user account and group information
and has also been updated to load the session against the database information
by the stored token on the given account.

A Common Class object with a function for debugging has been added to help
with testing the session and authentication portion of the workflow.

A basic working model currently exists, however will still need work to 
handle group levels and/or permissions and signout is still not binding
correctly in JavaScript, but working on authority->delete() correctly.